backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
transformers==4.35.0
torch==2.1.0
accelerate==0.24.0
bitsandbytes==0.41.1
unsloth @ git+https://github.com/unslothai/unsloth.git
python-dotenv==1.0.0
sentencepiece==0.1.99
protobuf==4.25.1
einops==0.7.0
safetensors==0.4.1
runpod==1.3.0 